web-animations/animation-timeline/animation-timeline-get-animation-players-sort-order-test.html

<!DOCTYPE html>
<title>Test sort order of the sequence returned by AnimationTimeline.getAnimationPlayers() method</title>
<meta name="assert" content="The list returned by AnimationTimeline.getAnimationPlayers() is sorted in increasing order by player sequence number">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#the-animationtimeline-interface">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var doc = newHTMLDocument();
    var animationTarget = doc.createElement('div');
    var player1 = doc.timeline.play(animationTarget.animate(null));
    var player2 = doc.timeline.play(animationTarget.animate(null));
    var player3 = doc.timeline.play(animationTarget.animate(null));

    assert_array_equals(document.timeline.getAnimationPlayers(), [player1, player2, player3],
        'Set of objects returned by AnimationTimeline.getAnimationPlayers() method ' +
        'should be ordered by player sequence number');
}, 'Test set of objects returned by AnimationTimeline.getAnimationPlayers() method ' +
    'is ordered by player sequence number');


test(function() {
    var doc = newHTMLDocument();
    var animationTarget = doc.createElement('div');
    doc.body.appendChild(animationTarget);
    var player1 = doc.timeline.play(animationTarget.animate(null));
    var player2 = doc.timeline.play(animationTarget.animate(null));

    var doc2 = newHTMLDocument();
    var animationTarget2 = doc2.createElement('div');
    doc2.body.appendChild(animationTarget);
    var player3 = doc2.timeline.play(animationTarget2.animate(null));
    var player4 = doc2.timeline.play(animationTarget2.animate(null));

    assert_array_equals(doc.timeline.getAnimationPlayers(), [player1, player2],
        'Set of objects returned by AnimationTimeline.getAnimationPlayers() method ' +
        'should be ordered by player sequence number');
    assert_array_equals(doc2.timeline.getAnimationPlayers(), [player3, player4],
        'Set of objects returned by AnimationTimeline.getAnimationPlayers() method ' +
        'should be ordered by player sequence number');
}, 'Test set of objects returned by AnimationTimeline.getAnimationPlayers() method ' +
    'of different timelines is ordered by player sequence number');
</script>
</body>
